+++ app/components/component-docs/dataDisplay/CoreBadge.docs.js
@@ -1,14 +1,15 @@
 import {
   CoreClasses,
   CoreStack,
-  CoreSpan,
   CoreH4,
   CoreTypographyBody1,
   CoreIcon,
   CoreBadge,
-  CoreBox
+  CoreBox,
+  __IconTypes
 } from "@wrappid/core";
 
+import CodeBlock from "../../CodeBlock";
 import CodeImport from "../../CodeImport";
 import CodeSample from "../../CodeSample";
 import ComponentProps from "../../ComponentProps";
@@ -33,14 +34,27 @@ export default function CoreBadgeDocs() {
           "Examples of badges containing text, using primary and secondary\
             colors. The badge is applied to its children."
         }
-        code={`
+        code={`<CoreBox>
 <CoreBadge badgeContent={4} color="primary">
-  <CoreIcon color="action" icon="mail" />
-</CoreBadge>`}
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+</CoreBox>`}
         renderElement={
           <CoreBox>
             <CoreBadge badgeContent={4} color="primary">
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" color="action" icon="mail" />
+
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
           </CoreBox>
         }
@@ -50,27 +64,51 @@ export default function CoreBadgeDocs() {
         title={"Color"}
         description={
           <>
-            Use
-            <CoreSpan code={true}>{" color "}</CoreSpan>
-            prop to apply theme palette to component.
+            Use <CodeBlock>color</CodeBlock> prop to apply theme palette to component.
           </>
         }
-        code={`
-<CoreBadge badgeContent={4} color="secondary">
-  <CoreIcon color="action" icon="mail" />
-</CoreBadge>
-<CoreBadge badgeContent={4} color="success">
-  <CoreIcon color="action" icon="mail" />
-</CoreBadge>
+        code={`<CoreStack spacing={2} direction="row">
+  <CoreBadge badgeContent={4} color="secondary">
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+
+  <CoreBadge badgeContent={4} color="success">
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+</CoreStack>
         `}
         renderElement={
           <CoreStack spacing={2} direction="row">
             <CoreBadge badgeContent={4} color="secondary">
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" color="action" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
 
             <CoreBadge badgeContent={4} color="success">
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" color="action" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
           </CoreStack>
         }
@@ -79,52 +117,78 @@ export default function CoreBadgeDocs() {
       <CodeSample title="Customization (NOT IMPLEMENTED)" />
 
       <CodeSample title="Badge visibility (NOT IMPLEMENTED)" />
-        
-      <CodeSample
-        // title={"Badge visibility"}
-        // description={
-        //   <>
-        //     The visibility of badges can be controlled using the
-        //     {/* eslint-disable-next-line react/jsx-newline */}
-        //     <CoreSpan code={true}>{" invisible "}</CoreSpan>
-        //     prop.
-        //   </>
-        // }
-        // code={``}
-        // renderElement={<CoreStack spacing={2} direction="row"></CoreStack>}
-      />
-
+    
       <CodeSample
         title={"Maximum value"}
         description={
           <>
-            You can use the
-            <CoreSpan code={true}>{" max "}</CoreSpan> prop to cap the value of
+            You can use the{" "}
+
+            <CodeBlock>max</CodeBlock> prop to cap the value of
             the badge content.
           </>
         }
-        code={`
-<CoreBadge color="secondary" badgeContent={99}>
-    <CoreIcon icon="mail" />
-</CoreBadge>
-<CoreBadge color="secondary" badgeContent={100}>
-    <CoreIcon icon="mail" />
-</CoreBadge>
-<CoreBadge color="secondary" badgeContent={1000} max={999}>
-    <CoreIcon icon="mail" />
-</CoreBadge>`}
+        code={`<CoreStack spacing={2} direction="row">
+  <CoreBadge color="secondary" badgeContent={99}>
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+
+  <CoreBadge color="secondary" badgeContent={100}>
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+
+  <CoreBadge color="secondary" badgeContent={1000} max={999}>
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+</CoreStack>`}
         renderElement={
           <CoreStack spacing={2} direction="row">
             <CoreBadge color="secondary" badgeContent={99}>
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
 
             <CoreBadge color="secondary" badgeContent={100}>
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
 
             <CoreBadge color="secondary" badgeContent={1000} max={999}>
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
           </CoreStack>
         }
@@ -134,55 +198,46 @@ export default function CoreBadgeDocs() {
         title={"Dot badge"}
         description={
           <>
-            The
-            <CoreSpan code={true}>{" dot "}</CoreSpan>
+            The <CodeBlock>dot</CodeBlock>{" "}
             prop changes a badge into a small dot. This can be used as a notification that something has changed without giving a count.
           </>
         }
-        code={`
-<CoreBadge color="secondary" variant="dot">
-    <CoreIcon icon="mail" />
-</CoreBadge>`}
+        code={`<CoreStack spacing={2} direction="row">
+  <CoreBadge color="secondary" variant="dot">
+    <CoreIcon
+      type={__IconTypes.MATERIAL_OUTLINED_ICON}
+      color="primary"
+      fontSize= "medium"
+    >
+      mail
+    </CoreIcon>
+  </CoreBadge>
+</CoreStack>`}
         renderElement={
           <CoreStack spacing={2} direction="row">
             <CoreBadge color="secondary" variant="dot">
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
+              <CoreIcon
+                type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                color="primary"
+                fontSize= "medium"
+              >
+                mail
+              </CoreIcon>
             </CoreBadge>
           </CoreStack>
         }
       />
 
       <CodeSample
-        title={"Badge alignment"}
+        title="Badge alignment(NOT IMPLEMENTED)"
         description={
           <>
-            You can use the
-            <CoreSpan code={true}>{" anchorOrigin "}</CoreSpan>
-            prop to move the badge to any corner of the wrapped element.
+            You can use the <CodeBlock >anchorOrigin</CodeBlock>prop to move the badge to any corner of the wrapped element.
           </>
         }
         code={`
-<CoreBadge
-  anchorOrigin={{
-    vertical: 'top',
-    horizontal: 'right',
-    }}
->
-  <CoreIcon icon="mail" />
-</CoreBadge>
         `}
-        renderElement={
-          <CoreStack spacing={2} direction="row">
-            <CoreBadge
-              anchorOrigin={{
-                horizontal: "right",
-                vertical  : "top",
-              }}
-            >
-              <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
-            </CoreBadge>
-          </CoreStack>
-        }
+        renderElement={""}
       />
 
       <CodeSample
@@ -190,8 +245,7 @@ export default function CoreBadgeDocs() {
         description={
           <>
             You can&apos;t rely on the content of the badge to be announced
-            correctly. You should provide a full description, for instance, with
-            <CoreSpan code={true}>{" aria-label: "}</CoreSpan>
+            correctly. You should provide a full description, for instance, with <CodeBlock >aria-label</CodeBlock>
           </>
         }
         code={`
@@ -204,7 +258,13 @@ export default function CoreBadgeDocs() {
           <CoreStack spacing={2} direction="row">
             <CoreIcon aria-label={100}>
               <CoreBadge badgeContent={100} color="secondary">
-                <CoreIcon type="__IconTypes.FONTAWESOME_V5_SOLID_ICON" icon="mail" />
+                <CoreIcon
+                  type={__IconTypes.MATERIAL_OUTLINED_ICON}
+                  color="primary"
+                  fontSize= "medium"
+                > 
+                mail
+                </CoreIcon>
               </CoreBadge>
             </CoreIcon>
           </CoreStack>